feat: finish migrate(?)
