+++ site/src/pages/ProfessorPage/index.tsx
@@ -12,8 +12,8 @@ import Error from '../../component/Error/Error';
 
 import { setProfessor } from '../../store/slices/popupSlice';
 import { useAppSelector, useAppDispatch } from '../../store/hooks';
-import { CourseWithTermsLookup, ProfessorGQLData } from '../../types/types';
-import { searchAPIResult } from '../../helpers/util';
+import { ProfessorGQLData } from '../../types/types';
+import { searchAPIResult, unionTerms } from '../../helpers/util';
 
 const ProfessorPage: FC = () => {
   const { id } = useParams<{ id: string }>();
@@ -34,35 +34,6 @@ const ProfessorPage: FC = () => {
       });
     }
   }, [dispatch, id]);
-
-  const unionTerms = (courseHistory: CourseWithTermsLookup) => {
-    // quarters mapped to the order of when they occur in the calendar year
-    const quartersOrdered: Record<string, string> = {
-      Winter: 'a',
-      Spring: 'b',
-      Summer1: 'c',
-      Summer2: 'd',
-      Summer10wk: 'e',
-      Fall: 'f',
-    };
-
-    // get array of arrays of term names
-    const allTerms = Object.values(courseHistory);
-
-    // flatten and take union of array
-    const union = [...new Set(allTerms.flatMap((term) => term.terms))];
-
-    // sort so that the most recent term appears first in the dropdown
-    union.sort((a, b) => {
-      const [yearA, qtrA]: string[] = a.split(' ');
-      const [yearB, qtrB]: string[] = b.split(' ');
-      // first compare years (descending)
-      // if years are equal, compare terms (most recent first)
-      return yearB.localeCompare(yearA) || quartersOrdered[qtrB].localeCompare(quartersOrdered[qtrA]);
-    });
-
-    return union;
-  };
 
   // if professor does not exists
   if (error) {
