+++ types/src/search.ts
@@ -2,7 +2,7 @@ import { paths } from './generated/anteater-api-types';
 import { PrerequisiteTree } from './course';
 
 type SearchAAPIResponseEntry =
-  paths['/v2/rest/search']['get']['responses'][200]['content']['application/json']['data'][number];
+  paths['/v2/rest/search']['get']['responses'][200]['content']['application/json']['data']['results'][number];
 
 type _SearchAAPICourseResponseEntry = Extract<SearchAAPIResponseEntry, { type: 'course' }>;
 
@@ -10,6 +10,7 @@ interface SearchAAPIResponseCourseEntry extends _SearchAAPICourseResponseEntry {
   prerequisiteTree: PrerequisiteTree;
 }
 
-export type SearchAAPIResponse = Array<
-  Extract<SearchAAPIResponseEntry, { type: 'instructor' } | SearchAAPIResponseCourseEntry>
->;
+export type SearchAAPIResponse = {
+  count: number;
+  results: Array<Extract<SearchAAPIResponseEntry, { type: 'instructor' } | SearchAAPIResponseCourseEntry>>;
+};
